PESB-1 Refactor paquetes y clases
