+++ src/main/java/com/eoi/springboot/MySpringBootApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.eoi.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
 @RestController // Prepárate para recibir peticiones HTTP (REST)
-public class DemoApplication {
+public class MySpringBootApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(MySpringBootApplication.class, args);
 	}
 
 	@RequestMapping("/")
